requirements.txt
discord
datetime
python-dotenv
pynacl
interactions.py
